feat: break generate commands to submodules
The single generate cmd module will grow with a new code. This commits
splits it into separate module per generate. Each generator has now its
own module.
