+++ src/mobster/cmd/generate/modelcar.py
@@ -1,0 +1,23 @@
+"""A module for generating SBOM documents for OCI index images."""
+
+import logging
+from typing import Any
+
+from mobster.cmd.generate.base import GenerateCommand
+
+LOGGER = logging.getLogger(__name__)
+
+
+class GenerateModelcarCommand(GenerateCommand):
+    """
+    Command to generate an SBOM document for a model car task.
+    """
+
+    async def execute(self) -> Any:
+        """
+        Generate an SBOM document for modelcar.
+        """
+        # Placeholder for the actual implementation
+        LOGGER.debug("Generating SBOM document for modelcar")
+        self._content = {}
+        return self.content
