+++ src/mobster/cmd/generate/oci_artifact.py
@@ -1,0 +1,23 @@
+"""A module for generating SBOM documents for OCI artifact images."""
+
+import logging
+from typing import Any
+
+from mobster.cmd.generate.base import GenerateCommand
+
+LOGGER = logging.getLogger(__name__)
+
+
+class GenerateOciArtifactCommand(GenerateCommand):
+    """
+    Command to generate an SBOM document for an OCI artifact.
+    """
+
+    async def execute(self) -> Any:
+        """
+        Generate an SBOM document for OCI artifact.
+        """
+        # Placeholder for the actual implementation
+        LOGGER.debug("Generating SBOM document for OCI artifact")
+        self._content = {}
+        return self.content
