+++ src/mobster/cmd/generate/oci_index.py
@@ -1,8 +1,7 @@
-"""A command execution module for generating SBOM documents."""
+"""A module for generating SBOM documents for OCI index images."""
 
 import json
 import logging
-from abc import ABC
 from datetime import datetime, timezone
 from typing import Any
 from uuid import uuid4
@@ -19,50 +18,10 @@ from spdx_tools.spdx.model.relationship import Relationship, RelationshipType
 from spdx_tools.spdx.model.spdx_no_assertion import SpdxNoAssertion
 from spdx_tools.spdx.writer.write_anything import write_file
 
-from mobster.cmd.base import Command
+from mobster.cmd.generate.base import GenerateCommand
 from mobster.image import Image
 
 LOGGER = logging.getLogger(__name__)
-
-
-class GenerateCommand(Command, ABC):
-    """A base class for generating SBOM documents command."""
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-
-        self._content: Any = None
-
-    @property
-    def content(self) -> Any:
-        """
-        Get the content of the SBOM document.
-        """
-        return self._content
-
-    async def save(self) -> None:
-        """
-        Save the SBOM document to a file if the output argument is provided.
-        """
-        if self.cli_args.output:
-            LOGGER.debug("Saving SBOM document to '%s'", self.cli_args.output)
-            with open(self.cli_args.output, "w", encoding="utf8") as output_file:
-                json.dump(self.content, output_file, indent=2)
-
-
-class GenerateOciImageCommand(GenerateCommand):
-    """
-    Command to generate an SBOM document for an OCI image.
-    """
-
-    async def execute(self) -> Any:
-        """
-        Generate an SBOM document for OCI image.
-        """
-        # Placeholder for the actual implementation
-        LOGGER.debug("Generating SBOM document for OCI image")
-        self._content = {}
-        return self.content
 
 
 class GenerateOciIndexCommand(GenerateCommand):
@@ -269,48 +228,3 @@ class GenerateOciIndexCommand(GenerateCommand):
                 str(self.cli_args.output),
                 validate=True,
             )
-
-
-class GenerateProductCommand(GenerateCommand):
-    """
-    Command to generate an SBOM document for a product level.
-    """
-
-    async def execute(self) -> Any:
-        """
-        Generate an SBOM document for product.
-        """
-        # Placeholder for the actual implementation
-        LOGGER.debug("Generating SBOM document for product")
-        self._content = {}
-        return self.content
-
-
-class GenerateModelcarCommand(GenerateCommand):
-    """
-    Command to generate an SBOM document for a model car task.
-    """
-
-    async def execute(self) -> Any:
-        """
-        Generate an SBOM document for modelcar.
-        """
-        # Placeholder for the actual implementation
-        LOGGER.debug("Generating SBOM document for modelcar")
-        self._content = {}
-        return self.content
-
-
-class GenerateOciArtifactCommand(GenerateCommand):
-    """
-    Command to generate an SBOM document for an OCI artifact.
-    """
-
-    async def execute(self) -> Any:
-        """
-        Generate an SBOM document for OCI artifact.
-        """
-        # Placeholder for the actual implementation
-        LOGGER.debug("Generating SBOM document for OCI artifact")
-        self._content = {}
-        return self.content
