+++ src/mobster/cmd/generate/product.py
@@ -1,0 +1,23 @@
+"""A module for generating SBOM documents for products."""
+
+import logging
+from typing import Any
+
+from mobster.cmd.generate.base import GenerateCommand
+
+LOGGER = logging.getLogger(__name__)
+
+
+class GenerateProductCommand(GenerateCommand):
+    """
+    Command to generate an SBOM document for a product level.
+    """
+
+    async def execute(self) -> Any:
+        """
+        Generate an SBOM document for product.
+        """
+        # Placeholder for the actual implementation
+        LOGGER.debug("Generating SBOM document for product")
+        self._content = {}
+        return self.content
