+++ tests/cmd/generate/test_modelcar.py
@@ -1,0 +1,12 @@
+from unittest.mock import MagicMock
+
+import pytest
+
+from mobster.cmd.generate.modelcar import GenerateModelcarCommand
+
+
+@pytest.mark.asyncio
+async def test_GenerateModelcarCommand_execute() -> None:
+    command = GenerateModelcarCommand(MagicMock())
+
+    assert await command.execute() == {}
